Prototype for pickled static data
