+++ worlds/lingo/rules.py
@@ -1,8 +1,9 @@
 from typing import TYPE_CHECKING
 
 from BaseClasses import CollectionState
+from .datatypes import RoomAndDoor
 from .player_logic import AccessRequirements, LingoPlayerLogic, PlayerLocation
-from .static_logic import PROGRESSION_BY_ROOM, PROGRESSIVE_ITEMS, RoomAndDoor
+from .static_logic import PROGRESSION_BY_ROOM, PROGRESSIVE_ITEMS
 
 if TYPE_CHECKING:
     from . import LingoWorld
